♻️ Refactor(보드): userId AuthenticationUser교체

+++ src/main/java/com/wolf/workflow/board/service/BoardService.java
@@ -4,7 +4,6 @@ import com.wolf.workflow.board.adapter.BoardAdapter;
 import com.wolf.workflow.board.adapter.BoardUserAdapter;
 import com.wolf.workflow.board.dto.request.BoardRequestDto;
 import com.wolf.workflow.board.dto.request.BoardUpdateRequestDto;
-import com.wolf.workflow.board.dto.response.AssigneeResponseDto;
 import com.wolf.workflow.board.dto.response.BoardGetResponseDto;
 import com.wolf.workflow.board.dto.response.BoardResponseDto;
 import com.wolf.workflow.board.dto.response.BoardUpdateResponseDto;
@@ -14,6 +13,7 @@ import com.wolf.workflow.board.entity.BoardUserRole;
 import com.wolf.workflow.board.entity.InvitationStatus;
 import com.wolf.workflow.board.entity.Participation;
 import com.wolf.workflow.common.exception.NotFoundBoardException;
+import com.wolf.workflow.common.security.AuthenticationUser;
 import com.wolf.workflow.user.adapter.UserAdapter;
 import com.wolf.workflow.user.entity.User;
 import lombok.RequiredArgsConstructor;
@@ -23,9 +23,6 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-
-import java.util.List;
-import java.util.stream.Collectors;
 
 
 @Service
@@ -45,11 +42,10 @@ public class BoardService {
    * @throws NotFoundBoardException 보드를 찾을 수 없는 경우
    */
 
-  public BoardResponseDto createBoard(BoardRequestDto requestDto, Long userId) {
-    User user = userAdapter.getUserById(requestDto.getUserId());
-    Board board = Board.createBoard(requestDto, user);
+  public BoardResponseDto createBoard(AuthenticationUser user, BoardRequestDto requestDto) {
+    User findUser = userAdapter.getUserById(user.getUser().getId());
+    Board board = Board.createBoard(requestDto, findUser);
     boardAdapter.createBoard(board);
-
     return BoardResponseDto.of(board);
   }
 
@@ -61,12 +57,11 @@ public class BoardService {
    * @throws NotFoundBoardException 보드를 찾을 수 없는 경우
    */
 
-  public BoardUpdateResponseDto updateBoard(BoardUpdateRequestDto requestDto, Long boardId, Long userId) {
-    User user = userAdapter.getUserById(userId);
-    // 보드 존재 여부 체크
+  public BoardUpdateResponseDto updateBoard(AuthenticationUser user, BoardUpdateRequestDto requestDto, Long boardId) {
+    boardAdapter.checkManagerAuthority(user);
+    User findUser = userAdapter.getUserById(user.getUser().getId());
     Board board = boardAdapter.getBoardById(boardId);
-    board.updateBoard(requestDto.getBoard_name(), requestDto.getContent(), user.getId());
-
+    board.updateBoard(requestDto.getBoard_name(), requestDto.getContent(), findUser.getId());
     return BoardUpdateResponseDto.of(board);
   }
 
@@ -76,10 +71,9 @@ public class BoardService {
    * @param boardId
    * @throws NotFoundBoardException 보드를 찾을 수 없는 경우
    */
-  public void deleteBoard(Long boardId) {
-    // 보드 존재 여부 체크
-    boardAdapter.getBoardById(boardId);
-    // 보드 삭제
+  public void deleteBoard(AuthenticationUser user, Long boardId) {
+    boardAdapter.checkManagerAuthority(user);
+    boardAdapter.getBoardById(boardId); // 보드 존재 여부 체크
     boardAdapter.deleteBoard(boardId);
   }
 
@@ -126,17 +120,14 @@ public class BoardService {
    * @throws NotFoundBoardException 보드를 찾을 수 없는 경우
    */
 
-  public void inviteUserToBoard(Long boardId, Long userId) {
+  public void inviteUserToBoard(AuthenticationUser user, Long boardId, Long userId) {
+    boardAdapter.checkManagerAuthority(user);
     // 초대하려는 보드 존재 여부 체크
-    Board board = boardAdapter.getBoardById(boardId);
-
-    // 초대하려는 사용자 존재 여부 체크
-    User user = userAdapter.getUserById(userId);
-    // 초대하려는 사용자가 이미 보드에 초대된 경우 체크
-    boardUserAdapter.BoardUserExists(boardId, userId);
-
+    Board board = boardAdapter.getBoardById(boardId); // 보드 존재 여부 체크
+    User invitee = userAdapter.getUserById(userId); // 초대할 사용자 존재 여부 체크
+    boardUserAdapter.BoardUserExists(boardId, userId); // 이미 초대된 사용자 체크
     BoardUser newBoardUser = BoardUser.createBoardUser(
-        board, user, Participation.ENABLE, BoardUserRole.ASSIGNEE, InvitationStatus.ACCEPTED
+        board, invitee, Participation.ENABLE, BoardUserRole.ASSIGNEE, InvitationStatus.ACCEPTED
     );
     boardUserAdapter.saveBoardUser(newBoardUser);
   }
@@ -152,25 +143,8 @@ public class BoardService {
   public void approveInvitation(Long boardId, Long userId, boolean approve) {
     boardUserAdapter.approveInvitation(boardId, userId, approve);
   }
-
-  /**
-   *  보드 유저리스트 찾기
-   *
-   * @param boardId
-   * @return List<AssigneeResponseDto>
-   */
-  public List<AssigneeResponseDto> getAssigneesByBoardId(Long boardId) {
-
-    // 보드 존재 여부 체크
-    Board board = boardAdapter.getBoardById(boardId);
-
-    // 보드에 속한 BoardUser 리스트 가져오기
-    List<BoardUser> boardUsers = boardUserAdapter.getBoardUsersByBoardId(boardId);
-
-    // BoardUser 리스트를 AssigneeResponseDto 리스트로 변환
-    return boardUsers.stream()
-            .map(boardUser -> new AssigneeResponseDto(boardUser.getUser().getId(), boardUser.getUser().getNickName()))
-            .collect(Collectors.toList());
-  }
 }
+
+
+
